pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.isort]
combine_as_imports = "true"
include_trailing_comma = "true"
line_length = 120
multi_line_output = 3

[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "bug_tracker"
package-mode = false
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
celery = {extras = ["redis"], version = "^5.4.0"}
dj-database-url = "^2.3.0"
dj-rest-auth = {extras = ["with-social"], version = "^7.0.1"}
django = {extras = ["argon2"], version = "^5.1.6"}
django-axes = {extras = ["ipware"], version = "^7.0.1"}
django-filter = "^24.3"
django-storages = {extras = ["google"], version = "^1.14.5"}
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
drf-spectacular = "^0.28.0"
drf-spectacular-sidecar = "^2025.2.1"
flower = "^2.0.1"
markdown = "^3.7"
mysqlclient = "^2.2.7"
nh3 = "^0.2.21"
psycopg2-binary = "^2.9.10"
python = "^3.12"
python-decouple = "^3.8"
python-magic = "^0.4.27"

[tool.poetry.group.dev.dependencies]
bandit = "^1.8.2"
black = "^25.1.0"
djlint = "^1.36.4"
flake8 = "^7.1.1"
isort = "^6.0.0"
pre-commit = "^4.1.0"
pytest-django = "^4.10.0"
pytest-mock = "^3.14.0"
python-magic-bin = "^0.4.14"
ruff = "^0.9.4"
toml-sort = "^0.24.2"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "bug_tracker.settings.development"
markers = [
  "e2e",
  "integration",
  "unit"
]
python_files = ["*_tests.py", "test_*.py", "tests.py"]

[tool.ruff.lint.per-file-ignores]
"**/bug_tracker/settings/**.py" = ["F401", "F403"]
